Updated DEV.NIX to run in IDX, Fixed child-key expo warning, Renamed Files
