+++ .idx/dev.nix
@@ -3,52 +3,45 @@
 { pkgs, ... }: {
   # Which nixpkgs channel to use.
   channel = "stable-23.11"; # or "unstable"
-
   # Use https://search.nixos.org/packages to find packages
-  packages = [
-    # pkgs.go
-    # pkgs.python311
-    # pkgs.python311Packages.pip
-    # pkgs.nodejs_20
-    # pkgs.nodePackages.nodemon
-  ];
-
+  packages = [pkgs.nodejs_20];
   # Sets environment variables in the workspace
-  env = {};
+  env = {
+    EXPO_USE_FAST_RESOLVER = 1; 
+  };
   idx = {
     # Search for the extensions you want on https://open-vsx.org/ and use "publisher.id"
     extensions = [
-      # "vscodevim.vim"
+      "msjsdiag.vscode-react-native"
     ];
-
-    # Enable previews
+    workspace = {
+      # Runs when a workspace is first created with this `dev.nix` file
+      onCreate = {
+        install = "npm ci --prefer-offline --no-audit --no-progress --timing && npm i @expo/ngrok@^4.1.0";
+      };
+      # Runs when a workspace restarted
+      onStart = {
+        connect-device = ''
+          adb -s localhost:5554 wait-for-device
+        '';
+        android = ''
+          npm run android -- --port 5554 --tunnel
+        '';
+      };
+    };
+    # Enable previews and customize configuration
     previews = {
       enable = true;
       previews = {
-        # web = {
-        #   # Example: run "npm run dev" with PORT set to IDX's defined port for previews,
-        #   # and show it in IDX's web preview panel
-        #   command = ["npm" "run" "dev"];
-        #   manager = "web";
-        #   env = {
-        #     # Environment variables to set for your server
-        #     PORT = "$PORT";
-        #   };
-        # };
-      };
-    };
-
-    # Workspace lifecycle hooks
-    workspace = {
-      # Runs when a workspace is first created
-      onCreate = {
-        # Example: install JS dependencies from NPM
-        # npm-install = "npm install";
-      };
-      # Runs when the workspace is (re)started
-      onStart = {
-        # Example: start a background task to watch and re-build backend code
-        # watch-backend = "npm run watch-backend";
+        web = {
+          command = ["npm" "run" "web" "--" "--port" "$PORT"];
+          manager = "web";
+        };
+        android = {
+          # noop
+          command = ["tail" "-f" "/dev/null"];
+          manager = "web";
+        };
       };
     };
   };
